refactor: move mappers to mapper folder
