+++ src/main/java/br/db/ecotrack/ecotrack_api/service/UserService.java
@@ -1,17 +1,17 @@
 package br.db.ecotrack.ecotrack_api.service;
 
 import java.util.Optional;
+
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import br.db.ecotrack.ecotrack_api.controller.request.UserRequestDto;
 import br.db.ecotrack.ecotrack_api.controller.response.UserResponseDto;
 import br.db.ecotrack.ecotrack_api.domain.entity.User;
-import br.db.ecotrack.ecotrack_api.domain.mapper.UserMapper;
+import br.db.ecotrack.ecotrack_api.mapper.UserMapper;
 import br.db.ecotrack.ecotrack_api.repository.UserRepository;
 import jakarta.persistence.EntityNotFoundException;
-
-import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class UserService {
